refactor: Move package to pishing (#70)
