+++ frontend/app/src/main/java/com/example/vishingguard/pishing/vishing/data/VishingResponse.kt
@@ -1,0 +1,11 @@
+package com.example.vishingguard.pishing.vishing.data
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class VishingResponse(
+    @SerialName("status") val status: Int,
+    @SerialName("message") val message: String,
+    @SerialName("data") val data: Boolean?
+)
